requirements.txt
Flask==2.3.2
line-bot-sdk==3.6.0
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-auth-oauthlib==1.2.2
firebase-admin==6.3.0
requests==2.31.0
beautifulsoup4==4.12.0
pydantic==2.11.3
PyQt6==6.9.0
PyQt6-WebEngine==6.9.0
plotly==5.19.0
dotenv==0.9.9